Move WrappedComparisonGrid and WrappedFeatureGrid into their own individual directories.
